learn-pytorch/learn-ddp/archive/full-examples/requirements.txt
hydra-core==1.1.1
pytorch==1.10.0
tqdm==4.62.3
torchmetrics==0.6.0
torchvision==0.11.1
